refactor(instagram-embed): extract url/css utils so tests can skip the handler
Drop the nitropack/runtime vitest alias + stub by moving the pure URL
rewriting + CSS scoping helpers out of the handler file. The handler
now re-imports them; the unit test imports directly from the utils
module and no longer pulls the `withSigning` → `nitropack/runtime`
virtual chain at load time.
