--- test/unit/__mocks__/nitropack-runtime.ts
@@ -1,6 +1,0 @@
-export function useRuntimeConfig() {
-  return { public: {} } as any
-}
-export function useNitroApp() {
-  return {} as any
-}
